Refactoring reaction message visualization code
