+++ ord_schema/visualization/template.txt
@@ -1,0 +1,191 @@
+To a 
+
+{# VESSEL #}
+{{ reaction.setup.vessel|vessel_prep }} 
+{{ reaction.setup.vessel.preparation_details|parenthetical_if_def }} 
+{{ reaction.setup.vessel|vessel_size }} 
+{{ reaction.setup.vessel|vessel_material }} 
+{{ reaction.setup.vessel.material_details|parenthetical_if_def }} 
+{{ reaction.setup.vessel|vessel_type }} 
+{{ reaction.setup.vessel.details|parenthetical_if_def }} 
+
+was 
+
+{# AUTOMATION #}
+{% if reaction.setup.is_automated %} automatically {% endif %}
+{{ reaction.setup.automation_platform|parenthetical_if_def }}
+
+added 
+
+{# REACTION INPUTS - SORT BY ADDITION ORDER #}
+{% for group in reaction.inputs.values()|groupby('addition_order') %}
+    {% if loop.last %} and {% endif %}
+
+    {% if reaction|uses_addition_order %}
+        ({{ group.list[0].addition_order }})
+    {% endif %}
+
+    {% for input in group.list %}
+        {% for compound in input.components %}
+
+            {# COMPOUND DETAILS #}
+            {{ compound|compound_amount }}
+            {{ compound|compound_name }} 
+            {{ compound|compound_role }} 
+            {{ compound|compound_source_prep }} 
+
+            {% if not loop.last %} + {% endif %}
+        {% endfor %}
+        {{ input|input_addition }}{% if not loop.last %}, {% endif %}
+    {% endfor %}
+
+    {% if not loop.last %}; {% endif %}
+{% endfor %}. 
+
+{# TEMPERATURE #}
+{{ reaction.conditions.temperature|temperature_conditions }} 
+
+{# PRESSURE #}
+{{ reaction.conditions.pressure|pressure_conditions }} 
+
+{# STIRRING #}
+{{ reaction.conditions.stirring|stirring_conditions }} 
+
+{# TODO(ccoley) ILLUMINATION #}
+
+{# TODO(ccoley) ELECTROCHEMISTRY #}
+
+{# TODO(ccoley) FLOW #}
+
+{# OBSERVATIONS - images not handled #}
+{% if reaction.observations %}
+    During the reaction, the following observations were noted: 
+    {% for observation in reaction.observations %}
+        {% if loop.last  and loop.index > 1 %} and {% endif %}
+        {% if observation.comment %}
+            after 
+            {% if observation.time.value %}
+                {{ observation.time|unit_format }}
+            {% else %}
+                an unspecified amount of time
+            {% endif %}, 
+            {{ observation.comment }}
+            {% if not loop.last %}; {% endif %}
+        {% endif %}
+    {% endfor %}.
+{% endif %}
+
+{# TODO(ccoley) WORKUP #}
+{% if reaction.workup %}
+    The workup procedure consisted of 
+    {% for workup in reaction.workup %}
+        {% if loop.last  and loop.index > 1 %} and {% endif %}
+        ({{ loop.index }}) 
+            <WORKUP_STEP>
+        {% if not loop.last %}; {% endif %}
+    {% endfor %}
+{% endif %}
+
+{# OUTCOME #}
+{% for outcome in reaction.outcomes %}
+    The reaction was {% if not loop.first %} also {% endif %} analyzed after 
+    {% if outcome.reaction_time.value %}
+        {{ outcome.reaction_time|unit_format }}
+    {% else %}
+        an unspecified amount of time
+    {% endif %} 
+    by  
+    {% for analysis in outcome.analyses.values() %}
+        {% if loop.last and loop.index > 1 %} and {% endif %}
+        {{ analysis|analysis_format }}
+        {{ analysis.details|parenthetical_if_def }}
+        {% if not loop.last %}, {% endif %}
+    {% endfor %}.
+
+    {# PRODUCTS #}
+    {% for product in outcome.products %}
+        {{ product.compound|compound_name }}
+        {% if product.compound.is_desired_product %}
+            , the desired product,
+        {% endif %}
+
+        {# PRODUCT IDENTITY #}
+        {% if product.analysis_identity %}
+            was identified through
+            {% for key in product.analysis_identity %}
+                {% if loop.last and loop.index > 1 %} and {% endif %}
+                {{ outcome.analyses[key]|analysis_format }}
+                {% if not loop.last %}, {% endif %}
+            {% endfor %}
+        {% else %}
+            was observed
+        {% endif %}
+        
+        {# PRODUCT YIELD #}
+        {% if product.compound_yield.value %}
+            with a yield of {{ product.compound_yield.value|round(3) }}%
+            {% if product.compound_yield.precision %}
+                (p/m {{ product.compound_yield.precision|round(3) }}%)
+            {% endif %}
+            {% if product.analysis_yield %}
+                (assayed through
+                {% for key in product.analysis_yield %}
+                    {% if loop.last and loop.index > 1 %} and {% endif %}
+                    {{ outcome.analyses[key]|analysis_format }}
+                    {% if not loop.last %}, {% endif %}
+                {% endfor %})
+            {% endif %}
+        {% endif %}
+
+        {# PRODUCT SELECTIVITY #}
+        {% if product.selectivity.value %}
+            with a selectivity of {{ product.selectivity.value|round(3) }}
+            {{ product.selectivity|selectivity_type }}
+            {% if product.selectivity.precision %}
+                (p/m {{ product.selectivity.precision|round(3) }})
+            {% endif %}
+            {% if product.analysis_selectivity %}
+                (assayed through
+                {% for key in product.analysis_selectivity %}
+                    {% if loop.last and loop.index > 1 %} and {% endif %}
+                    {{ outcome.analyses[key]|analysis_format }}
+                    {% if not loop.last %}, {% endif %}
+                {% endfor %})
+            {% endif %}
+        {% endif %}
+
+        {# PRODUCT PURITY #}
+        {% if product.purity.value %}
+            with a purity of {{ product.purity.value|round(3) }}%
+            {% if product.purity.precision %}
+                (p/m {{ product.purity.precision|round(3) }}%)
+            {% endif %}
+            {% if product.analysis_purity %}
+                (assayed through
+                {% for key in product.analysis_purity %}
+                    {% if loop.last and loop.index > 1 %} and {% endif %}
+                    {{ outcome.analyses[key]|analysis_format }}
+                    {% if not loop.last %}, {% endif %}
+                {% endfor %})
+            {% endif %}
+        {% endif %}.
+
+        {# PRODUCT COLOR/TEXTURE #}
+        {{ product|product_color_texture }}.
+
+
+    {% endfor %}
+
+{% endfor %}
+
+{# DICLAIMER ABOUT SCHEMA LIMITATIONS #}
+{% if reaction.conditions.conditions_are_dynamic %}
+    Note: the conditions for this reaction may not precisely fit the schema, so
+    additional care should be taken when reading this description.
+{% endif %}
+{% if reaction.conditions.details %}
+    Condition details: {{ reaction.conditions.details }}.
+{% endif %}
+{% if reaction.notes.procedure_details %}
+    Procedure details: {{ reaction.notes.procedure_details }}.
+{% endif %}
